Test getting-started-app
