+++ .luacheckrc
@@ -1,4 +1,4 @@
 include_files = {'**/*.lua', '*.luacheckrc', '*.rockspec', 'cartridge'}
-exclude_files = {'.rocks/', 'tmp/'}
+exclude_files = {'**/.rocks/', 'tmp/'}
 max_line_length = 120
 redefined = false
