+++ Makefile
@@ -25,13 +25,13 @@ lint: bootstrap
 	flake8
 
 .PHONY: test
-test: unit integration test-getting-started
+test: unit integration test-examples
 
 python_deps:
 	pip3 install -r test/python/requirements.txt
 
 .PHONY: integration
-integration: bootstrap
+integration:
 	python3 -m pytest test/integration
 
 .PHONY: unit
@@ -40,14 +40,9 @@ unit: bootstrap
 	.rocks/bin/luatest -v --coverage && .rocks/bin/luacov .
 	grep -A999 '^Summary' tmp/luacov.report.out
 
-.PHONY: test-getting-started
-test-getting-started: bootstrap
-	cd test/examples/getting-started-app; \
-		sh test_start.sh ../../../examples/getting-started-app;
-	cd ./examples/getting-started-app; \
-		.rocks/bin/luatest -v
-	.rocks/bin/luacheck ./examples/getting-started-app \
-		--exclude-files **/.rocks/*
+.PHONY: test-examples
+test-examples:
+	python3 -m pytest test/examples
 
 .PHONY: ci_prepare
 ci_prepare: python_deps
