+++ test/integration/test_version.py
@@ -4,6 +4,6 @@ from utils import run_command_and_get_output
 
 
 def test_version_command(cartridge_cmd):
-    rc, output = run_command_and_get_output([cartridge_cmd, "version"])
+    rc, output = run_command_and_get_output([cartridge_cmd, "--version"])
     assert rc == 0
     assert 'Tarantool Cartridge CLI v' in output
